feat: implement modal window with bug
